refactor: Reorganize files and folder structure
