+++ src/components/sections/About.tsx
@@ -1,4 +1,4 @@
-import SectionTitle from "./common/SectionTitle";
+import SectionTitle from "../ui/SectionTitle";
 import { profileInfo } from "@/lib/site";
 
 export default function About() {
